requirements.txt
httpx[http2]
python-dateutil
pydantic>=2.0.0
flask>=3.0.0
pytz>=2023.3
openai>=1.35.0
pytest-asyncio>=0.21.0
nylas>=6.0.0
python-dotenv>=1.0.0
colorama>=0.4.6